Реализовал родительский класс Component
