+++ src/task.js
@@ -1,7 +1,8 @@
-import {createElement} from '../src/utils.js';
+import {Component} from "./component";
 
-export class Task {
+export class Task extends Component {
   constructor(data) {
+    super();
     this._title = data.title;
     this._dueDate = data.dueDate;
     this._tags = data.tags;
@@ -10,6 +11,7 @@ export class Task {
     this._repeatingDays = data.repeatingDays;
     this._state = {};
     this._onEdit = null;
+    this._onEditBtnClick = this._onEditBtnClick.bind(this);
   }
 
 
@@ -21,10 +23,6 @@ export class Task {
     if (typeof this._onEdit === `function`) {
       this._onEdit();
     }
-  }
-
-  get element() {
-    return this._element;
   }
 
   set onEdit(fn) {
@@ -125,21 +123,10 @@ export class Task {
 
   bind() {
     this._element.querySelector(`.card__btn--edit`)
-      .addEventListener(`click`, this._onEditBtnClick.bind(this));
+      .addEventListener(`click`, this._onEditBtnClick);
   }
 
   unbind() {
-    this._element.removeEventListener(`click`, this._onEditBtnClick.bind(this));
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
+    this._element.removeEventListener(`click`, this._onEditBtnClick);
   }
 }
